The starting phase of an Argot Java/Js demo
This sets up a web socket server in a Tomcat container that can be connected to, each connection will receive data at a regular interval. Currently this is just stock data.

The front end html has been pulled in from the pure js demo (https://github.com/argotsdk/argot-demo) but is not wired up to functional javascript yet.

Pulling the js over as is isn't really appropriate because of the Socket.io js calling back to a known place on the node.js server that we don't have here. It looks like we'll be adding bower next.